UAR-896 - Moifying qualifications to allow added aggregator to recall
Added logic to RoleServiceImpl to change qualification that a
Principal only needs to have role 1109 (Proposal Aggregator),
and be associated with the given proposal number.

The new logic checks that the principal is associated by name
or group, and that the role is still active.

This logic lives in its own private method, and is called once
by principalHasRole(). If role 1109 tests as false in this method,
the logic will continue on as normal.