UAR-984 Correcting branch from development
We had to correct this branch when we did the merge to the release branch. The original UAR-984 was created from the development branch and there are commits in development that we do not want to go to release. This UAR-984-fixed branch was a branch that was created from master. We then coppied the files that were modified from UAR-984 branch to this branch.

Original comment from UAR-984:

created terminate method in KualiDocumentActionBase to use the disapprove functionality with a different confirmation